[ISSUE #3624]fix DefaultMQPushConsumerTest.testPullMessage_ExceptionOccursWhenComputePullFromWhere throws RejectedExecution
